+++ flutter/lib/src/native/c/sentry_native.dart
@@ -1,11 +1,14 @@
 import 'dart:async';
 import 'dart:ffi';
+import 'dart:io';
 import 'dart:typed_data';
 
+import 'package:collection/collection.dart';
 import 'package:ffi/ffi.dart';
 import 'package:meta/meta.dart';
 
 import '../../../sentry_flutter.dart';
+import '../../replay/replay_config.dart';
 import '../native_app_start.dart';
 import '../native_frames.dart';
 import '../sentry_native_binding.dart';
@@ -19,10 +22,18 @@ class SentryNative with SentryNativeSafeInvoker implements SentryNativeBinding {
   final SentryFlutterOptions options;
 
   @visibleForTesting
-  static final native = binding.SentryNative(DynamicLibrary.open('sentry.dll'));
+  static final native = binding.SentryNative(DynamicLibrary.open(
+      '$dynamicLibraryDirectory${Platform.isWindows ? 'sentry.dll' : 'libsentry.so'}'));
+
+  /// If the path is just the library name, the loader will look for it in
+  /// the usual places for shared libraries:
+  /// - on Linux in /lib and /usr/lib
+  /// - on Windows in the working directory and System32
+  @visibleForTesting
+  static String dynamicLibraryDirectory = '';
 
   @visibleForTesting
-  static String? crashpadPath;
+  static String? crashpadPath = _getDefaultCrashpadPath();
 
   SentryNative(this.options);
 
@@ -259,6 +270,14 @@ class SentryNative with SentryNativeSafeInvoker implements SentryNativeBinding {
   }
 
   @override
+  bool get supportsReplay => false;
+
+  @override
+  FutureOr<void> setReplayConfig(ReplayConfig config) {
+    _logNotSupported('replay config');
+  }
+
+  @override
   FutureOr<SentryId> captureReplay(bool isCrash) {
     _logNotSupported('capturing replay');
     return SentryId.empty();
@@ -386,4 +405,20 @@ extension on List<dynamic> {
     }
     return cObject;
   }
+}
+
+String? _getDefaultCrashpadPath() {
+  if (Platform.isLinux) {
+    final lastSeparator =
+        Platform.resolvedExecutable.lastIndexOf(Platform.pathSeparator);
+    if (lastSeparator >= 0) {
+      final appDir = Platform.resolvedExecutable.substring(0, lastSeparator);
+      final candidates = [
+        '$appDir${Platform.pathSeparator}crashpad_handler',
+        '$appDir${Platform.pathSeparator}bin/crashpad_handler'
+      ];
+      return candidates.firstWhereOrNull((path) => File(path).existsSync());
+    }
+  }
+  return null;
 }