+++ flutter/lib/src/native/sentry_native_channel.dart
@@ -7,6 +7,7 @@ import 'package:flutter/services.dart';
 import 'package:meta/meta.dart';
 
 import '../../sentry_flutter.dart';
+import '../replay/replay_config.dart';
 import 'native_app_start.dart';
 import 'native_frames.dart';
 import 'method_channel_helper.dart';
@@ -186,8 +187,15 @@ class SentryNativeChannel
   @override
   Future<List<DebugImage>?> loadDebugImages(SentryStackTrace stackTrace) =>
       tryCatchAsync('loadDebugImages', () async {
-        final images = await channel
-            .invokeListMethod<Map<dynamic, dynamic>>('loadImageList');
+        Set<String> instructionAddresses = {};
+        for (final frame in stackTrace.frames) {
+          if (frame.instructionAddr != null) {
+            instructionAddresses.add(frame.instructionAddr!);
+          }
+        }
+
+        final images = await channel.invokeListMethod<Map<dynamic, dynamic>>(
+            'loadImageList', instructionAddresses.toList());
         return images
             ?.map((e) => e.cast<String, dynamic>())
             .map(DebugImage.fromJson)
@@ -208,6 +216,18 @@ class SentryNativeChannel
 
   @override
   Future<void> nativeCrash() => channel.invokeMethod('nativeCrash');
+
+  @override
+  bool get supportsReplay => false;
+
+  @override
+  FutureOr<void> setReplayConfig(ReplayConfig config) =>
+      channel.invokeMethod('setReplayConfig', {
+        'width': config.width,
+        'height': config.height,
+        'frameRate': config.frameRate,
+        'bitRate': config.bitRate,
+      });
 
   @override
   Future<SentryId> captureReplay(bool isCrash) =>