+++ flutter/test/screenshot/masking_config_test.dart
@@ -1,8 +1,9 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
-import 'package:sentry_flutter/src/replay/masking_config.dart';
+import 'package:sentry_flutter/src/screenshot/masking_config.dart';
 
+import '../mocks.dart';
 import 'test_widget.dart';
 
 void main() async {
@@ -114,16 +115,15 @@ void main() async {
   });
 
   group('$SentryReplayOptions.buildMaskingConfig()', () {
-    List<String> rulesAsStrings(SentryReplayOptions options) {
-      final config = options.buildMaskingConfig();
+    List<String> rulesAsStrings(SentryPrivacyOptions options) {
+      final config =
+          options.buildMaskingConfig(MockLogger().call, PlatformChecker());
       return config.rules
           .map((rule) => rule.toString())
           // These normalize the string on VM & js & wasm:
           .map((str) => str.replaceAll(
-              RegExp(
-                  r"SentryMaskingDecision from:? [fF]unction '?_maskImagesExceptAssets[@(].*",
-                  dotAll: true),
-              'SentryMaskingDecision)'))
+              RegExp(r"=> SentryMaskingDecision from:? .*", dotAll: true),
+              '=> SentryMaskingDecision)'))
           .map((str) => str.replaceAll(
               ' from: (element, widget) => masking_config.SentryMaskingDecision.mask',
               ''))
@@ -131,39 +131,42 @@ void main() async {
     }
 
     test('defaults', () {
-      final sut = SentryReplayOptions();
+      final sut = SentryPrivacyOptions();
       expect(rulesAsStrings(sut), [
         ...alwaysEnabledRules,
         '$SentryMaskingCustomRule<$Image>(Closure: (Element, Widget) => SentryMaskingDecision)',
         '$SentryMaskingConstantRule<$Text>(mask)',
-        '$SentryMaskingConstantRule<$EditableText>(mask)'
+        '$SentryMaskingConstantRule<$EditableText>(mask)',
+        '$SentryMaskingCustomRule<$Widget>(Closure: ($Element, $Widget) => $SentryMaskingDecision)'
       ]);
     });
 
     test('maskAllImages=true & maskAssetImages=true', () {
-      final sut = SentryReplayOptions()
+      final sut = SentryPrivacyOptions()
         ..maskAllText = false
         ..maskAllImages = true
         ..maskAssetImages = true;
       expect(rulesAsStrings(sut), [
         ...alwaysEnabledRules,
         '$SentryMaskingConstantRule<$Image>(mask)',
+        '$SentryMaskingCustomRule<$Widget>(Closure: ($Element, $Widget) => $SentryMaskingDecision)'
       ]);
     });
 
     test('maskAllImages=true & maskAssetImages=false', () {
-      final sut = SentryReplayOptions()
+      final sut = SentryPrivacyOptions()
         ..maskAllText = false
         ..maskAllImages = true
         ..maskAssetImages = false;
       expect(rulesAsStrings(sut), [
         ...alwaysEnabledRules,
         '$SentryMaskingCustomRule<$Image>(Closure: (Element, Widget) => SentryMaskingDecision)',
+        '$SentryMaskingCustomRule<$Widget>(Closure: ($Element, $Widget) => $SentryMaskingDecision)'
       ]);
     });
 
     test('maskAllText=true', () {
-      final sut = SentryReplayOptions()
+      final sut = SentryPrivacyOptions()
         ..maskAllText = true
         ..maskAllImages = false
         ..maskAssetImages = false;
@@ -171,15 +174,19 @@ void main() async {
         ...alwaysEnabledRules,
         '$SentryMaskingConstantRule<$Text>(mask)',
         '$SentryMaskingConstantRule<$EditableText>(mask)',
+        '$SentryMaskingCustomRule<$Widget>(Closure: ($Element, $Widget) => $SentryMaskingDecision)'
       ]);
     });
 
     test('maskAllText=false', () {
-      final sut = SentryReplayOptions()
+      final sut = SentryPrivacyOptions()
         ..maskAllText = false
         ..maskAllImages = false
         ..maskAssetImages = false;
-      expect(rulesAsStrings(sut), alwaysEnabledRules);
+      expect(rulesAsStrings(sut), [
+        ...alwaysEnabledRules,
+        '$SentryMaskingCustomRule<$Widget>(Closure: ($Element, $Widget) => $SentryMaskingDecision)'
+      ]);
     });
 
     group('user rules', () {
@@ -187,22 +194,23 @@ void main() async {
         ...alwaysEnabledRules,
         '$SentryMaskingCustomRule<$Image>(Closure: (Element, Widget) => SentryMaskingDecision)',
         '$SentryMaskingConstantRule<$Text>(mask)',
-        '$SentryMaskingConstantRule<$EditableText>(mask)'
+        '$SentryMaskingConstantRule<$EditableText>(mask)',
+        '$SentryMaskingCustomRule<$Widget>(Closure: ($Element, $Widget) => $SentryMaskingDecision)'
       ];
       test('mask() takes precedence', () {
-        final sut = SentryReplayOptions();
+        final sut = SentryPrivacyOptions();
         sut.mask<Image>();
         expect(rulesAsStrings(sut),
             ['$SentryMaskingConstantRule<$Image>(mask)', ...defaultRules]);
       });
       test('unmask() takes precedence', () {
-        final sut = SentryReplayOptions();
+        final sut = SentryPrivacyOptions();
         sut.unmask<Image>();
         expect(rulesAsStrings(sut),
             ['$SentryMaskingConstantRule<$Image>(unmask)', ...defaultRules]);
       });
       test('are ordered in the call order', () {
-        var sut = SentryReplayOptions();
+        var sut = SentryPrivacyOptions();
         sut.mask<Image>();
         sut.unmask<Image>();
         expect(rulesAsStrings(sut), [
@@ -210,7 +218,7 @@ void main() async {
           '$SentryMaskingConstantRule<$Image>(unmask)',
           ...defaultRules
         ]);
-        sut = SentryReplayOptions();
+        sut = SentryPrivacyOptions();
         sut.unmask<Image>();
         sut.mask<Image>();
         expect(rulesAsStrings(sut), [
@@ -218,7 +226,7 @@ void main() async {
           '$SentryMaskingConstantRule<$Image>(mask)',
           ...defaultRules
         ]);
-        sut = SentryReplayOptions();
+        sut = SentryPrivacyOptions();
         sut.unmask<Image>();
         sut.maskCallback(
             (Element element, Image widget) => SentryMaskingDecision.mask);
@@ -231,7 +239,7 @@ void main() async {
         ]);
       });
       test('maskCallback() takes precedence', () {
-        final sut = SentryReplayOptions();
+        final sut = SentryPrivacyOptions();
         sut.maskCallback(
             (Element element, Image widget) => SentryMaskingDecision.mask);
         expect(rulesAsStrings(sut), [
@@ -240,7 +248,7 @@ void main() async {
         ]);
       });
       test('User cannot add $SentryMask and $SentryUnmask rules', () {
-        final sut = SentryReplayOptions();
+        final sut = SentryPrivacyOptions();
         expect(sut.mask<SentryMask>, throwsA(isA<AssertionError>()));
         expect(sut.mask<SentryUnmask>, throwsA(isA<AssertionError>()));
         expect(sut.unmask<SentryMask>, throwsA(isA<AssertionError>()));